requirements.txt
asgiref==3.4.1; python_version >= "3.6"
django-cors-headers==3.10.0; python_version >= "3.6"
django==3.2.9; python_version >= "3.6"
djangorestframework-simplejwt==5.0.0; python_version >= "3.7"
djangorestframework==3.12.4; python_version >= "3.5"
gunicorn==20.1.0; python_version >= "3.5"
psycopg2-binary==2.9.2; python_version >= "3.6"
pyjwt==2.3.0; python_version >= "3.7"
pytz==2021.3; python_version >= "3.6"
sqlparse==0.4.2; python_version >= "3.6"
whitenoise==5.3.0; python_version >= "3.5" and python_version < "4"
